requirements.txt
websocket-client==0.59.0
websockets
